Корректное переименование файлов
