+++ View/CircuitMainForm.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace View
+{
+    public partial class MainForm : Form
+    {
+        public MainForm()
+        {
+            InitializeComponent();
+        }
+    }
+}
