+++ View/CircuitMainForm.resx
