+++ View/ImpedanceForm.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace View
+{
+    public partial class ImpedanceForm : Form
+    {
+        public ImpedanceForm()
+        {
+            InitializeComponent();
+        }
+    }
+}
